Changed i into an unsigned long int.

+++ 0x19-hash_tables/6-hash_table_delete.c
@@ -9,7 +9,7 @@
 
 void hash_table_delete(hash_table_t *ht)
 {
-	unsigned int i = 0;
+	unsigned long int i = 0;
 	hash_node_t **array, *temp, *current;
 
 	array = ht->array;
